requirements.txt
chess>=1.10
edq-utils>=0.0.5
